fix: Resolved comments [#123]
